Assembler: Throw parse errors on unrecognised instructions and operands
One of the biggest annoyances with the assembler is it's tendency to
ignore bad assembly, either dropping instructions or just producing a
corrupt binary that crashes DirectX.

This patch makes the assembler throw parse error exceptions if it
encounters an instruction or operand that it cannot parse, which should
go a long way towards alleviating bugs caused by typos.

It also catches the exception back up in the assembler() routine to add
the failing line number to the exception message before throwing it on
to the caller.