Assembler: Skip lines containing only whitespace
Now that the assembler throws exceptions on unrecognised instructions we
have to be careful not to get it to barf on some innocuous comment. It
already has code to skip lines that are either blank or start with a
comment as their first character, and to strip comments from the end of
lines, but it is missing the case where a comment is indented on an
otherwise blank line.

This changes the blank line test from checking for an empty string to
checking that there is some non-space character in the string. At this
point comments have already been stripped and tabs have been converted
to spaces, so this should also treat a line that contains nothing but a
comment as blank, regardless of indentation.

With this change every assembly shader in my 3d-fixes repository
assembles without throwing any parse error exceptions, which is a pretty
good indication that it should work well for others.