TestShaders: Add test cases from HLSLCrossCompiler
HLSLCrossCompiler has a collection of test cases with very good coverage
of the shader instruction set, so these can also be useful as test cases
for us.

This commit just adds the binary files from their test cases since this
is the most useful format for both our assembler and HLSL test cases.
The upstream repository also contains HLSL source and assembly for all
of these, which may be a useful reference, but doesn't really need to be
in our repository.

vs_2_0, vs_3_0, ps_2_0 and ps_3_0 shaders have been excluded as
irrelevant.

A following commit will hook these up to our test suite.