TestShaders: Hook up HLSLCrossCompiler shaders to assembler test suite
This adds all the HLSLCrossCompiler test cases to our assembler test
suite. Almost all are passing, with the following exceptions:

    FAIL: BinaryDecompiler/apps/shaders/shaders/SubroutinesPS.o...
    FAIL: BinaryDecompiler/ps5/interfaces.o...
    FAIL: BinaryDecompiler/ps5/interfaces_multifunc.o...
    FAIL: BinaryDecompiler/ps5/interface_arrays.o...
    FAIL: BinaryDecompiler/ps5/retc.o...
    FAIL: BinaryDecompiler/quarantined/ps5/this.o...

This is pretty much expected, as functions, class linkage, etc. are not
yet supported by the assembler on account of never encountering them in
the wild nor being able to coax fxc to generate a shader containing
them, and indeed these all fail with "Unrecognised instruction:
dcl_function_body fb0"