Assembler: Add support for printf/errorf debugging instructions
These two instructions can be used when the debug layer is enabled to
retrieve arbitrary output from shaders, and may prove useful to us if we
can find a way for shaderhackers to see their output (e.g. perhaps by
enabling the debug layer from the DirectX control panel).

This adds parsing support for these instructions to the assembler and
disassembler, which are inserted into the shader as a custom data field
using opcode 0x35.

The debug test cases are now passing.