Mark d3dx_user.ini dirty if it contains any invalid settings
Invalid settings can end up in the d3dx.ini when a user removes, renames
or potentially upgrades a mod that had saved persistent settings.
3DMigoto will warn about these on every config load until they are
removed when the config is next saved, but since the config is only
saved after a persistent setting has been altered this may not happen
for some time. The end result is a lot of people asking how to fix these
errors.

We don't necessarily want to suppress them outright though, because they
might raise a red flag if something is wrong. This aims to compromise
between the two by flagging the user config as dirty immediately if it
contains any invalid settings so that it will be updated on the next
config reload or clean game shutdown, and displays a message on the
overlay indicating this and informing the user of how to deal with it,
and most of the usual "Unrecognised entry" warnings are suppressed with
only the first showing up to provide some context.