Dump IIDs on DX object creation & include ReShade + SpecialK IIDs
The analyse_iunknown() function is useful to get some info about what
kind of DirectX object we are dealing with, but currently we only use it
if we aren't able to find our HackerDevice to find out why.

It is also potentially useful in other situations, such as to find out
what DirectX interfaces are supported or if there are any third party
tools in play, or if the COM identity rule has been violated, so just
call it whenever we create a DX device / context / swap chain.

This also adds the ReShade and SpecialK IIDs to this function, so that
it can report if one of those tools are sitting between us and DirectX
since we have had known compatibility issues with those tools in the
past. Also add the 3DMigoto IIDs for completeness and to alert us if we
ever get into a situation where we have somehow double hooked in.