Treat assembler parse errors as non-fatal in some circumstances
Since we never used to throw exceptions on bad assembly in the past it
is likely that there will be some fixes in the wild that have done
something wrong that will be flagged by the new exceptions. e.g. I've
seen at least one case of a shaderhacker including the comment block
lines we use to include the assembly inside a dumped HLSL shader in an
assembly shader, which will fail to parse and now prevent the
replacement shader from being used at all.

This commit allows the assembler to treat parse errors as non-fatal in
some cases - specifically when a shader is replaced at the time of
shader creation, and when the ShaderRegex engine patches a shader. These
cases will display the parse error on the overlay if hunting is enabled,
and remain silent in release mode so as not to bug users who don't know
how to deal with it.

The shader timestamp is not recorded in the creation case to ensure that
F10 will attempt to reload the shader until the shaderhacker fixes the
issue. The shader cache will also be disabled for these shaders (though
this cannot be done retrospectively if the shaderhacker already had
their cache enabled before upgrading 3DMigoto).

However, we explicitly do not allow parse errors if a shader is either
being reloaded live or from cmd_Decompiler. In these cases the
shaderhacker at the keyboard should be able to deal with their bug.

Note that compared to previous behaviour this will drop a bad
instruction that triggers a parse error, whereas previously it may have
partially assembled the instruction - meaning there is still the
possibility for a behaviour change in some edge cases. However, cases
where the assembler partially assembled an instruction before tended to
result in crashes inside DirectX or the driver, so skipping the
instruction is likely going to be a better option anyway.