Fix AB-BA type deadlock between Overlay, DirectX and the release tracker
A hang was observed in Resident Evil 2. Launced with hunting=2, started
a new game, ran/looked around the starting area for a few seconds (until
the "look around" tutorial prompt was on screen) then pressed F10 and
the game hung.

This is a race condition, so it only hit some of the time, and if it
didn't hit the first time the overlay was displayed it would not hit at
all on account of DirectXTK caching the object it created on demand the
first time.

I see two threads currently inside 3DMigoto - the first is in the
ResourceReleaseTracker waiting for our critical section (the first clue
that there is a locking bug present):

    ntdll.dll!NtWaitForAlertByThreadId()	Unknown
    ntdll.dll!RtlpWaitOnAddressWithTimeout()	Unknown
    ntdll.dll!RtlpWaitOnCriticalSection()	Unknown
    ntdll.dll!RtlpEnterCriticalSectionContended()	Unknown
    ntdll.dll!RtlEnterCriticalSection()	Unknown
    d3d11.dll!ResourceReleaseTracker::Release() Line 1279	C++
    d3d11.dll!std::list<struct std::pair<struct SGuid const ,class CPrivateData>,class std::allocator<struct std::pair<struct SGuid const ,class CPrivateData> > >::clear(void)	Unknown
    d3d11.dll!CResource<ID3D11Texture2D1>::~CResource<ID3D11Texture2D1>()	Unknown
    d3d11.dll!NDXGI::CDeviceChild<IDXGIResource1,IDXGISwapChainInternal>::FinalRelease()	Unknown
    d3d11.dll!CUseCountedObject<NOutermost::CDeviceChild>::UCDestroy()	Unknown
    d3d11.dll!CUseCountedObject<NOutermost::CDeviceChild>::UCReleaseUse()	Unknown
    d3d11.dll!NDXGI::CDeviceChild<IDXGISurface,IUnknown>::FinalRelease()	Unknown
    d3d11.dll!CLayeredObjectWithCLS<CShaderResourceView>::CContainedObject::Release()	Unknown
    re2.exe!0000000150374dd0()	Unknown
    re2.exe!000000014ffabc53()	Unknown
    re2.exe!000000014ffef8b8()	Unknown
    re2.exe!000000014f10e41f()	Unknown
    kernel32.dll!BaseThreadInitThunk()	Unknown
    ntdll.dll!RtlUserThreadStart()	Unknown

The second thread in 3DMigoto is currently in the overlay with our
critical section held, but is not doing anything that we would naively
expect would block the other thread - it is currently trying to draw the
overlay notifications via DirectXTK:

    ntdll.dll!NtWaitForAlertByThreadId()	Unknown
    ntdll.dll!RtlpWaitOnAddressWithTimeout()	Unknown
    ntdll.dll!RtlpWaitOnCriticalSection()	Unknown
    ntdll.dll!RtlpEnterCriticalSectionContended()	Unknown
    ntdll.dll!RtlEnterCriticalSection()	Unknown
    d3d11.dll!CDevice::CreateBlendState_Worker()	Unknown
    [Inline Frame] d3d11.dll!DirectX::CommonStates::Impl::CreateBlendState(D3D11_BLEND) Line 85	C++
    [Inline Frame] d3d11.dll!DirectX::CommonStates::AlphaBlend::__l2::<lambda_2f151d2d4e66d961bb9354ffe71cf4a5>::operator()(ID3D11BlendState * *) Line 216	C++
    [Inline Frame] d3d11.dll!DirectX::DemandCreate(Microsoft::WRL::ComPtr<ID3D11BlendState> &) Line 39	C++
    d3d11.dll!DirectX::CommonStates::AlphaBlend() Line 214	C++
    d3d11.dll!Overlay::DrawRectangle(float x, float y, float w, float h, float r, float g, float b, float opacity) Line 487	C++
    d3d11.dll!Overlay::DrawNotices(float * y) Line 706	C++
    d3d11.dll!Overlay::DrawOverlay() Line 806	C++
    d3d11.dll!HackerSwapChain::RunFrameActions() Line 278	C++
    d3d11.dll!HackerSwapChain::Present(unsigned int SyncInterval, unsigned int Flags) Line 555	C++
    re2.exe!0000000150383b18()	Unknown
    re2.exe!0000000150006340()	Unknown
    re2.exe!000000015000425a()	Unknown
    re2.exe!000000014ef64c08()	Unknown
    re2.exe!000000014f10e41f()	Unknown
    kernel32.dll!BaseThreadInitThunk()	Unknown
    ntdll.dll!RtlUserThreadStart()	Unknown

We have an AB-BA lock ordering bug here (confirmed by checking the
DirectX critical section memory to identify which thread currently owns
it), with the two general orderings being:

1. We take our critical section, then call into a DirectX routine
   (CreateBlendState in this case, but there are likely many others)
   which takes their lock

2. The game releases a resource dropping it's refcount to 0. DirectX
   takes their lock to do whatever cleanup they need to do then calls
   Release() on any interfaces associated with the resource via
   SetPrivateDataInterface, which calls our release tracker. Our release
   tracker attempts to take our critical section to allow it to safely
   purge any references of the resource from our data structures.

= AB-BA type Deadlock!

This specific case is a race condition where one thread calls Present()
while another thread simultaneously releases a resource.

To resolve this specific instance of this deadlock we are removing the
use of G->mCriticalSection from the overlay, replacing it with a more
fine grained lock that specifically protects the notices data structure.
This replaces the AB-BA locking graph with AB-CA, which has no cycles
and therefore no deadlock.

Note that this depends on DrawOverlay() never being called with the
G->mCriticalSection held, so warnings to this effect have been added to
several key areas in the code. Ideally we would also tag the routines
with this constraint to be verified through either static code analysis
or runtime checks, but I am unaware of what (if any) solutions exist to
do this on Windows (lockdep is the solution used in the Linux kernel)

Note that this only looks at one confirmed AB-BA deadlock scenario, and
there could potentially be other deadlock scenarios if we ever call into
DirectX with the G->mCriticalSection lock held, which will require a
code audit to uncover.