Move CriticalSection wrapper class into util.h
This class is useful to wrap a CRITICAL_SECTION structure to take care
of initializing and destroying it in a C++ constructor to reduce boiler
plate. It may be useful for more than just the notices lock going
forward, so move it into util.h so it can be used wherever we need.