Add locking around surface creation mode and resource creation
The NvAPI_Stereo_SetSurfaceCreationMode routine affects resource
creation globally for the device, so it is not thread nor context safe.
Most games limit their resource creation to a single thread making this
a non-issue, but games like RE2/RE7/DMC5 and DOA6 are known to create
resources on whatever thread suits their mood.

This can lead to a race if we ever force the surface creation mode,
since while we have the creation mode set another thread could create a
resource and unexpectedly get a mono (or stereo) resource, potentially
even leaving the final surface creation mode in the wrong state. I
suspect this may be the cause of some render targets or entire games
inexplicably switching to 2D, and have suspect it may even affect
StereoParams itself in some instances.

To attempt to address this issue, this patch adds a lock that must be
taken whenever we are overriding the surface creation mode, and whenever
we are creating a resource (with or without setting the surface creation
mode). This should serialise the resource creation to eliminate this
race - of course, being a rarely hit race condition this will need
further testing to gain confidence that this is the case and that this
is the correct fix.

Since this lock must be taken in nvstereo.h when StereoParams is
updated, and this header is not located in the DX11 project there are
some complications with using the lock debugging version of
EnterCriticalSectionPretty (the lock debugging depends on TLS and
Deviare, so it can't be split out of the DX11 project until they are) it
adds some macros specifically for this lock that will use the debug
facilities if they are available (if lock.h has been included in
advance) and the regular EnterCriticalSection if not.