Work around Present vs Release race condition in Resident Evil 2
Resident Evil 2 (and 7) are notorious for hanging in 3D Vision. I
managed to catch one of these in a debugger and determined that this is
a race between a resource being Release()d in one thread while another
thread simultaneously calls Present(). In this scenario the Present()
thread holds the DirectX lock, and the Release() thread is blocked
waiting for it, while the Present() thread is blocked waiting on a
handle to be signaled and is unable to make forward progress (this
appears to be a live lock rather than a deadlock, as the Present()
thread seems to continually retry).

Stack trace of the Release() thread:

    ntdll.dll!NtWaitForAlertByThreadId()    Unknown
    ntdll.dll!RtlpWaitOnAddressWithTimeout()    Unknown
    ntdll.dll!RtlpWaitOnCriticalSection()    Unknown
    ntdll.dll!RtlpEnterCriticalSectionContended()    Unknown
    ntdll.dll!RtlEnterCriticalSection()    Unknown
    d3d11.dll!CLayeredObjectWithCLS<CShaderResourceView>::CContainedObject::Release()    Unknown
    re2.exe!0000000150374dd0()    Unknown
    re2.exe!000000014ffabc53()    Unknown
    re2.exe!000000014ffef8b8()    Unknown
    re2.exe!000000014f10e41f()    Unknown
    kernel32.dll!BaseThreadInitThunk()    Unknown
    ntdll.dll!RtlUserThreadStart()    Unknown

Stack trace of the Present() thread:

    ntdll.dll!NtWaitForSingleObject()    Unknown
    KernelBase.dll!WaitForSingleObjectEx()    Unknown
    nvwgf2umx.dll!00007ffa79c85aa6()    Unknown
    nvwgf2umx.dll!00007ffa79c869d0()    Unknown
    nvwgf2umx.dll!00007ffa79aeb6ff()    Unknown
    nvwgf2umx.dll!00007ffa79aee188()    Unknown
    nvwgf2umx.dll!00007ffa79ca08bd()    Unknown
    nvwgf2umx.dll!00007ffa79b3c159()    Unknown
    d3d11.dll!NDXGI::CDevice::PresentImpl()    Unknown
    d3d11.dll!NDXGI::CDevice::Present()    Unknown
    dxgi.dll!CFlipPresentToDWM<struct CDXGISwapChainWrapper>::FlipPresentCore(void)    Unknown
    dxgi.dll!CDXGISwapChain::FlipPresentToDWM(struct SPresentArgsCore<struct CDXGISwapChainWrapper> const *,unsigned int,unsigned int,unsigned int &,unsigned int,struct tagRECT const *,struct DXGI_SCROLL_RECT const *,struct DXGI_INTERNAL_CONTENT_PROTECTION const &)    Unknown
    dxgi.dll!CDXGISwapChain::PresentImplCore(struct SPresentArgsCore<struct CDXGISwapChainWrapper> const *,unsigned int,struct tagRECT const *,unsigned int,struct DXGI_SCROLL_RECT const *,struct IDXGIResource *,struct CDXGISwapChain::PresentResultActions &)    Unknown
    dxgi.dll!CDXGISwapChain::PresentImpl()    Unknown
    dxgi.dll!CDXGISwapChain::Present()    Unknown
    GameOverlayRenderer64.dll!00007ffa7daba503()    Unknown
    d3d11.dll!HackerSwapChain::Present(unsigned int SyncInterval, unsigned int Flags) Line 567    C++
    re2.exe!0000000150383b18()    Unknown
    re2.exe!0000000150006340()    Unknown
    re2.exe!000000015000425a()    Unknown
    re2.exe!000000014ef64c08()    Unknown
    re2.exe!000000014f10e41f()    Unknown
    kernel32.dll!BaseThreadInitThunk()    Unknown
    ntdll.dll!RtlUserThreadStart()    Unknown

This bug is reminiscent of the race we recently uncovered between
3DMigoto's overlay and a resource being simultaneously released in
another thread (see #104), however in this case 3DMigoto is not at fault
since our lock is not held by any thread when this live lock occurs (we
are not even in the call stack of the Release() thread). It is not clear
if the 3D Vision driver is partially to blame or not - forcing a
StereoMode=2 seems to exacerbate the issue, but that may just be
enlarging a race window that already existed - i.e. this might purely be
a bug in Resident Evil 2 & 7 that boils down to violating the (poorly
documented) threading constraints of the DirectX 11 API.

Adding some debug to check which thread is doing what shows that for the
most part the resource releases do happen from the rendering thread as
they should, however every now and then this will be intermixed with
resources being released from a separate thread - these are the risky
ones. This is a tally of which threads are used for the Present() call
vs releasing resources in a very short run of RE2:

   count thread task
   ----- ----- -----
    2472 0d24: Running frame actions.  Device: 0000000008500CB0
     838 0d24: ResourceReleaseTracker::Release
      51 0d5c: Running frame actions.  Device: 0000000008500CB0
    1219 0d5c: ResourceReleaseTracker::Release
     521 184c: ResourceReleaseTracker::Release
       6 3780: Running frame actions.  Device: 0000000008500CB0
       5 3780: ResourceReleaseTracker::Release
      93 3f44: ResourceReleaseTracker::Release
    1908 3fa4: ResourceReleaseTracker::Release
    2182 6170: ResourceReleaseTracker::Release

While this is a game bug, 3DMigoto may be able to work around the issue.
We do have an opportunity to hold onto an extra reference to the object
that is to be Released(), preventing it from being freed and hopefully
avoiding the DirectX lock from being taken in that thread. We can then
periodically scan over all objects we are holding this reference to and
locate any with a refcount of exactly 1 indicating that we are the last
remaining reference holder and release them at that time from a thread
that is safe to do so.

This commit is the first attempt at implementing this workaround. The
approach used here is to tie this workaround into [TextureOverride]s -
this is done because reports are that it is textures forced to mono via
StereoMode=2 that end up triggering or exacerbating this hang, so this
first attempt allows the workaround to be enabled at the same time we
force the stereo mode:

    [TextureOverride_General_Rain]
    match_type = Buffer
    match_byte_width = > 90000
    match_bind_flags = 0x88
    match_misc_flags = 0x40
    match_stride = 220
>>> StereoMode = 2
>>> workaround_release_present_race = true
    match_priority = 1

That said, I am not yet convinced that these are the only resources that
can trigger this live lock, and we may end up needing to apply the
workaround to every resource. While we can make a match all
TextureOverride to handle that, if that does turn out to be the case it
would be simpler (and probably more appropriate) to turn it into a
global option instead.

For now, we will try this approach and see if the hang is eliminated or
if it can still be reproduced.