First pass at installing our own crash handler
This is a first attempt at implementing our own crash handler to make
our lives a bit easier on crashes and hangs. It is currently enabled by
adding "crash=1" into the [Logging] section and will install a global
exception handler via SetUnhandledExceptionFilter.

It also spawns a thread to monitor for a Ctrl+Alt+F12 keyboard salute,
which will raise an exception allowing us to trigger the exception
handler by hand in the event of a deadlock / hang.

When the exception handler is triggered it will:
- Sound an SOS
- Flush the log file to give us as much context as possible leading up
  to the crash
- Log the exception info
- Create a minidump file
- Attempt to switch the swap chain to windowed mode
- Wait for a debugger to be attached, triggering a debugbreak when one
  appears and continuing the program
- Exit the game if escape is pressed

This is currently fairly rough functionality - in particular the switch
to windowed mode seems a bit hit or miss.

At present this does not use set_terminate() to catch unhandled C++
exceptions.