Rework Present/Release race fix into a global option
With the previous commit in place to work around the Release() vs
Present() deadlock for specific resources (and applied to all
StereoMode=2 overrides) we still occasionally see a hang between these
calls. This would seem to fit with the theory that this issue may not be
related to any property of specific resources and may just apply more
generally to the bad threading behaviour of RE2/7.

There is some documentation of the DirectX threading constraints here:

    https://docs.microsoft.com/en-gb/windows/desktop/direct3ddxgi/d3d10-graphics-programming-guide-dxgi#multithread-considerations

This commit reworks the workaround to remove the [TextureOverride]
option, instead replacing it with a global option under [Device], that
will keep all textures alive until the following present call when
enabled.

This rework will add a considerable amount of work to the Present()
call, which may not be great for performance - but there are still some
related driver crashes to examine before worrying about optimising that.