Include resource views in the release/present race workaround
Despite now bumping the reference count on every resource created by the
game, in some driver crash debugging sessions I noticed another thread
seemed to still be directly releasing a resource from the game and was
blocked waiting on the DirectX lock:

    d3d11.dll!CLayeredObjectWithCLS<CShaderResourceView>::CContainedObject::Release()

Since I don't have anything more than symbols to work with I do have to
guess at exactly what this call is doing, and had assumed it was
releasing the contained resource inside the view - but that resource
should still have a reference from the current iteration of the
workaround. It may instead indicate that the resource view object is the
one being released here and trying to take the DirectX lock.

Notably, this was a crash debugging session rather than a deadlock
debugging session so the theory that these may contribute to hangs is
currently theoretical - but seems likely given the observed call stacks.

This commit adds the four types of resource views to the list of objects
that we will delay the release on. It may also turn out that we need to
also track all DeviceChild objects in the same way, but I have not yet
seen any other types of objects implicated in a debugging session.