Release/Present Workaround: Check 1/60th of the list each frame
Since we now bump the refcount on every single resource created by the
game our data structure of resources we need to manually check the
refcounts on can become quite large and iterating over the entire list
is a bit wasteful.

This aims to give us a reasonable compromise by only checking a subset
of the list each frame, so typically it may take a resource up to a
second to be released - which should curb vram usage from being able to
get too high, while still reducing the performance impact of the
workaround.

This lowers the present cost by over 400us, saving around 1fps on my
system and making it nearly indistinguishable as having the workaround
disabled (as measured using the 3DMigoto performance monitor).

In the future if we ever wrap the resources or hook their release calls
directly we could look at other more efficient solutions, but given this
workaround is only for one or two games it doesn't matter too much as
yet.