Remove documentation of the present/release workaround
While this option does seem to be quite effective at eliminating the
hangs seen in Resident Evil 2 from threading bugs, they are
unfortunately replaced by crashes inside the driver, meaning that this
option is not really ready for prime-time (arguably a crash is easier
for an end user to deal with than a hang since they don't have to use
task manager which may be obscured if they don't have it set to always
on top, but neither are good).

At the moment I don't have a whole lot to go off in investigating these
crashes - they happen in arbitrary driver threads when nvwgf2umx.dll
dereferences an invalid pointer (Not usually NULL, but another clearly
invalid value like 0x0000000000000005).

The timing of these and the fact that they happen in this game strongly
suggests that they likely have a similar cause to the hangs, but at this
point I do not believe that they are (directly) related.

This game is unusual in that it is constantly creating & destroying
resources (and does not restrict this to just the rendering thread),
which can exacerbate a number of bugs (as it already has for threading
bugs in 3DMigoto and the game itself, and would have also exacerbated
3DMigoto's previous resource mixup bugs if we hadn't already fixed
those) and my working hypothesis at the moment is that there is a
similar bug in the 3D Vision driver that it is also exacerbating.

If my hypothesis is correct, one potential strategy we could look at in
the future to mitigate this would be to add a resource pool, so that
whenever the game frees a resource it is added to the pool for reuse if
the game later attempts to create a new resource with a matching
description (and StereoMode). For this to work we would need to
intercept the Release call in a different way, such as by wrapping the
resource object or hooking the Release call.

If however my hypothesis is incorrect this would be a lot of work for
nothing, and at the moment I don't have any sort of smoking gun to give
me any significant confidence that my hypothesis is correct, so leaving
this for a future exercise when we might have wrapped the resource
objects for other reasons.