Make assembly sign off lines in HLSL comments in assembly as well
Some of our Shaderhackers copy the assembly comment block out of a HLSL
decompiled shader into a new assembly file if they decide to switch to
asm (as opposed to dumping a fresh assembly version or grabbing it from
ShaderCache), but at least a few of them missed removing the ~~~~~~~~*/
line that ends the C/HLSL style comment block, and that line will now be
treated as a parse error by the assembler.

This is not hugely critical, since the assembler now continues after
encountering parse errors on initial load, and future fixes will issue
the parse error alerting the Shaderhacker to remove that line, but we
can also make their workflow just a tad easier by changing that line so
that it is recognised as a comment by the assembler, like //~~~~~~~~~*/
- that way a shaderhacker can copy it straight into an assembly file
without having to worry about remembering to remove the line.

Obviously this change will not retrospectively affect already dumped
shaders.