Injector: Remove UAC UIAccess requirement (for now)
UAC UIAccess is required for injecting into Windows Store apps, but it
has the downside that the application needs to be digitally signed with
a key trusted by the system and will outright refuse to run if the
application is not signed.

The injector approach is also useful for other applications - and in
particular it allows us to get 3DMigoto into DOA6 (which otherwise exits
if it detects third party d3d11/dxgi DLLs present effectively blocking
the normal loading methods of 3DMigoto, ReShade and SpecialK). Disable
the UIAccess requirement so that the injector can be used without having
to deal with the signing requirement - once we have the rest of the
pieces in place needed for Windows Store Apps we can bring the setting
back, perhaps in a separate build configuration or via an external
manifest file.