Bail out of DllMain when injected into the wrong process
The DLL injection method used by injector.exe does not target any
specific executable and will inject 3DMigoto into every newly spawned
process in the system.

This is not really desirable - while usually quite harmless if the
d3dx.ini is missing, 3DMigoto can potentially interfere with the DirectX
API and other things, so we really don't want it active in any process
that we aren't actually targeting. Even when it is entirely harmless the
global injection is still problematic, because it locks the DLL making
it hard to deploy new versions during development, and preventing an end
user from uninstalling it.

To address this we make 3DMigoto check whether it is intended to be
running in the current process from DllMain, and will bail if it
determines that it not.

There are four distinct cases where 3DMigoto should continue running:

1. When loaded into the target process from the game directory
2. When loaded into the target process via the injector
3. When loaded into the injector process
4. When loaded into rundll32 as the profile helper

For cases 1 and 3 we can simply check if the DLL is located in the same
directory as the running executable. Case 4 we check explicitly,
matching both the executable name and the entry point listed on the
command line.

For case 2 we load the d3dx.ini from the injector directory, locate the
[Injector] section and compare the target setting with the executable
name. This check can be on just the executable name itself, or may
include one or more directories for cases where the executable name is
too generic to reliably match (e.g. "game.exe")

Since DllMain has limitations on what API calls we can make, and since
we don't want to spend any more time running in any undesired processes
than we want to. To this end, this adds a minimalistic "lite" ini parser
that uses nothing more than simple C pointer arithmetic and never
backtracks while searching for the target section & setting. The calling
routine is also restricted to Kernel32 APIs and standard library
routines that should be statically linked in. This lite ini parser is
added in a separate file so that it can also be used by the injector
project in a later commit.