Injector: Add support for ini config & begin polish
This adds support for parsing the [Injector] section in the injection
executable. It checks to make sure a target has been set and allows the
3DMigoto module name to be customised, which may be handy to tweak in
some cases as it can influence whether 3DMigoto needs to hook into the
d3d11 exports or not.

This adds some useful error messages to the injector if it is run
without being properly set up.