Fix whitespace errors in version information & update copyright
Several of the version string fields have rogue spaces before the NULL
terminator that should not be there, so remove them. The remaining
string fields using macros are correct as is since they are two separate
strings that will be concatenated together.

Bump the copyright year to 2019. We should probably set that one
automatically at build time because we are sure to forget to manually
bump it every year.