Injector: Verify 3DMigoto version
This checks to make sure that the module to be injected is 3DMigoto
v1.3.15 or later. We have had support for this injection method in
3DMigoto since v1.3.5, but without the check in DllMain to bail out of
unwanted processes the older versions are not considered safe and this
will stop them from being inadvertently injected by the unwary.

The version check can be disabled via "check_version = false" for those
who like to live dangerously, or to inject some other DLL that supports
the same injection technique, such as Helix Mod.