Bump 3DMigoto's refcount when injected
If the injector is closed too early things can go a little wrong:
- If it is closed very early we won't be loaded yet
- If it is closed after the game initialises DirectX everything is fine
- If it is closed somewhere between the above two then we might have
  been injected and hooked a bunch of functions, but then get
  prematurely unloaded when the injector is closed and potentially
  leading to a crash when the game does eventually go to initialise
  DirectX. I assume that the injector must be holding a reference on our
  module in the target process and once this goes away the module
  refcount drops to zero and gets unloaded prematurely.

To try to make things a bit more consistent, bump the refcount on our
module once we have verified that it is running inside the intended
target. That should prevent the module from being prematurely unloaded
so it is still around when the game gets onto initialising DirectX.

The benefit here is that we can get away with much shorter delays in the
injector.