Injector: Detect already running instances
This checks if an instance of the 3DMigoto loader is already running and
exits with an error message if it is. This just helps make sure things
remain a bit more predictable and to help make sure that the user won't
end up injecting multiple instances of 3DMigoto into every task
increasing the stall factor.