Work with files from DLL directory when injected
This sets 3DMigoto to work with files relative to the DLL directory
rather than the game's executable directory. Normally these will match
since 3DMigoto is usually loaded from the game's directory, but when
3DMigoto has been externally injected it will be in a different
location.

This allows the workflow of the injector to be "extract the zip file
anywhere you want and run the injector.exe" rather than requiring it be
extracted to the game directory (which should work anyway) and will be
necessary when we later add Windows Store app support to the injector
since we have no way to modify the game directory in that case.

This covers:
- d3dx.ini
- Included configuration files
- ShaderFixes
- ShaderCache
- d3d11_log.txt
- Custom Shaders
- Custom resources
- proxy_d3d11
- Frame analysis
- ShaderUsage.txt