Warn if d3dx.ini found in both DLL and EXE directories
We now use the d3dx.ini and other files from the DLL directory, but to
account for the fact that a user might have got confused and extracted
an older or newer version of the fix to the game directory but be
running the loader from a separate directory we will check to see if
both exist and alert the user of the possible issue.

Also log the game and DLL paths to make sure we can spot if a user is
loading 3DMigoto from outside the game directory or not.