Injector: Add require_admin option
If a game runs as admin, then the injector process will also need to run
as admin to be able to inject 3DMigoto. While it is fairly straight
forward to right click the injector and run as admin, it is easy to
forget and there isn't a good clear error message since the injector
doesn't get confirmation if the injection was actually successful.

This adds a require_admin option to the injector to have it re-launch
itself as admin if the user forgets.