Injector: Look for injected module in target process
This searches the modules in the target process to check if the injected
DLL is present as a more reliable way of determining when the injection
has completed successfully.

In many cases this removes the need for the delay, so make the delay
default to 0 seconds and make the wait_for_target undocumented since we
will almost always want it enabled.

Make the indefinite delay case keep looking for any more successful
injections since it is going to stay running.